+++ priv/templates/components/search_field.exs
@@ -21,6 +21,7 @@
       rounded: ["extra_small", "small", "medium", "large", "extra_large", "full"],
       space: ["extra_small", "small", "medium", "large", "extra_large"],
       only: ["search_field"],
+      helpers: [],
       module: ""
     ],
     optional: [],